src/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.7
bidict==0.23.1
certifi==2025.1.31
chainlit==2.1.2
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
duckduckgo_search==7.3.2
faiss-cpu==1.10.0
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
fsspec==2025.2.0
googleapis-common-protos==1.66.0
groq==0.18.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
joblib==1.4.2
Lazify==0.4.0
literalai==0.1.103
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx
numpy==2.2.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pillow==11.1.0
primp==0.12.0
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMuPDF==1.25.3
pypdf==5.2.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
syncer==2.0.3
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.29.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
